refactor(#5031): join lease types and heartbeat module
